siam_1812022_6

+++ Core/Entities/RptInvestmentSummary.cs
@@ -9,7 +9,7 @@ namespace Core.Entities
         public string ReferenceNo { get; set; }
         public string DonationTypeName { get; set; }
         public string DonationTo { get; set; }
-        public long ProposedAmount { get; set; }
+        public double ProposedAmount { get; set; }
         public DateTimeOffset FromDate { get; set; }
         public DateTimeOffset ToDate { get; set; }
         public string InvStatus { get; set; }
